+++ src/main/java/nenesekai/leetscope/entity/Assignment.java
@@ -10,8 +10,8 @@ import java.util.Date;
 @Data
 @TableName("`assignment`")
 public class Assignment {
-    private Long id;
-    private Long uid;
+    private Integer id;
+    private Integer uid;
     private String title;
     private String description;
     @TableField(fill = FieldFill.INSERT)
