+++ src/main/java/nenesekai/leetscope/service/SubmissionService.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Service;
 public interface SubmissionService {
     public Result create(Submission submission);
 
-    public Result listSubmissions(Long uid, Long assignmentId);
+    public Result listSubmissions(Integer uid, Integer assignmentId);
 }
